Create new class for input checking and move current input checking methods to here
