+++ src/vn/edu/rmit/kuri/Main.java
@@ -36,22 +36,6 @@ public class Main {
     System.out.printf("Database Path: %s\n", databasePath);
   }
 
-  public static boolean isIntegerFormat(String input) {
-    try {
-      Integer.parseInt(input);
-      return true;
-    } catch (NumberFormatException exception) {
-      return false;
-    }
-  }
-
-  public static void checkInput(String input, Scanner sc) {
-    while (!isIntegerFormat(input)) {
-      System.out.print("Your input is invalid. Please try again.\n>>>\s");
-      input = sc.nextLine();
-    }
-  }
-
   public static void main(String[] args) {
     String databasePath;
